Http Errors now throw exceptions
Earlier, http requests were returning a status code != 200 but cloud team couldn't find them because it wasn't in the error message.

Now error messages are sent by calling a runtimeexception with a custom error message.

Needed to change exception type to runtime so it didn't have to be declared to be thrown.

Also added HttpHandlerException to differentiate it from normal runtime exceptions.